backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
jinja2>=3.1.2
sqlalchemy>=2.0.0
face-recognition>=1.3.0
opencv-python>=4.7.0
pillow>=9.5.0
numpy>=1.24.0
aiofiles>=0.8.0
python-socketio>=5.8.0
pydantic-settings>=2.0.0
email-validator>=2.0.0 
passlib>=1.7.4
psycopg2-binary>=2.9.0

